refactored the design weights folder, into edits, derived and design weight (#703)
* refactored the design weights folder, into edits, derived and design weights

* Fixed issues with update column and clean postcode tests
