+++ tests/weights/test_calculate_generic_dweight_variables.py
@@ -2,7 +2,7 @@ import pyspark.sql.functions as F
 from chispa import assert_df_equality
 from pyspark.sql.window import Window
 
-from cishouseholds.weights.design_weights import calculate_combined_design_weights
+from cishouseholds.pipeline.design_weights import calculate_combined_design_weights
 
 
 def test_calculate_generic_design_weight_variables(spark_session):
