+++ tests/weights/test_calculate_scenario_c_antibody_dweights.py
@@ -1,6 +1,6 @@
 from chispa import assert_df_equality
 
-from cishouseholds.weights.design_weights import calculate_raw_antibody_design_weights_with_migration
+from cishouseholds.pipeline.design_weights import calculate_raw_antibody_design_weights_with_migration
 
 
 def test_calculate_scenario_c_antibody_design_weights(spark_session):
